Move submissions to own directory
- include urls in conferences
