+++ meetings/meetings/urls.py
@@ -15,9 +15,6 @@ urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'^users/(?P<pk>[0-9]+)/$', views.UserDetail, name='user-detail'),
     url(r'^conferences/', include('conferences.urls', namespace='conferences')),
-    url(r'^conferences/(?P<conference_id>[0-9]+)/submissions/$', views.SubmissionList.as_view(), name='submission-list'),
-    url(r'^conferences/(?P<conference_id>[0-9]+)/submissions/(?P<submission_id>[0-9]+)/$',
-        views.SubmissionDetail.as_view(), name='submission-detail'),
     url(r'^auth/', views.OsfAuthorizationUrl.as_view(), name='auth'),
     url(r'^login/', views.OsfAuthorizationCode.as_view(), name='login'),
     url(r'^admin/', admin.site.urls, name='admin'),
