+++ meetings/submissions/admin.py
@@ -1,4 +1,4 @@
 from django.contrib import admin
-from .models import Submission
+from submissions.models import Submission
 
 admin.site.register(Submission)
