+++ meetings/submissions/tests.py
@@ -1,0 +1,1 @@
+# TODO: Write lots of awesome tests
